refactor: move findCaretTokenIndex file to parser/common/utils
